Change SOD to precision 16, and change export to use float 16

+++ dataset/imagenet.py
@@ -223,7 +223,7 @@ class ImageNetDatasetExport(torchvision.datasets.ImageFolder):
         
 
         
-        np.save(sp_file_path, features)
+        np.save(sp_file_path, features.astype(np.float16))
         np.save(sp_file_path_target, np.array([target]))
 
         return torch.empty(0)
